chore: Update Docker build command in Makefile
This commit updates the Docker build command in the Makefile to remove the "--no-cache" flag. The "--no-cache" flag was previously used to ensure a clean build, but it is no longer necessary. Removing this flag will speed up the build process by allowing Docker to use cached layers when possible.