chore: Update .env.example and README.md files
This commit updates the .env.example file by removing the USER_LIMIT variable and updating the MAX_USERS_POPULATE value to 100. It also updates the README.md file by correcting the project description in both English and Portuguese. These changes improve the consistency and accuracy of the project configuration and documentation.